Fixed IndexOutOfBounds Exception while hitting the wall. Renamed files. Moved resources directory out of source code directory. Renamed directories. Further code cleaning needed.
