+++ Snake/src/sample/classes/Movement.java
@@ -1,8 +1,9 @@
-package sample.Class;
+package sample.classes;
 
 import javafx.scene.input.KeyCode;
 
 public enum Movement {
+    //TODO zmiana enumow to zla praktyka
     UP(KeyCode.W),
     DOWN(KeyCode.S),
     LEFT(KeyCode.A),
