+++ Snake/src/sample/classes/Sound.java
@@ -1,16 +1,17 @@
-package sample.Class;
+package sample.classes;
 
 import javafx.scene.media.AudioClip;
 import javafx.scene.media.Media;
 
 import java.nio.file.Paths;
-
+//TODO sciezka wzgledna
+//TODO resources -> obook src a nie w src
 public class Sound {
 
     private static AudioClip audioClip = null;
 
     public static void play(){
-        String path = "D:\\GitHubRepositories\\Java\\Snake\\src\\sample\\resources\\Soundtrack\\domkosnake44.mp3";
+        String path = "resources/soundtrack/domkosnake44.mp3";
         Media media = new Media(Paths.get(path).toUri().toString());
         audioClip = new AudioClip(media.getSource());
         audioClip.setVolume(0.20);
