+++ Snake/src/sample/fxml/GameController.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.text.Font?>
 
 <?import javafx.scene.control.Button?>
-<Pane fx:id="paneGame" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" onKeyReleased="#previousScene" prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: black;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.Controller.gameController">
+<Pane fx:id="paneGame" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" onKeyReleased="#previousScene" prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: black;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.controller.GameController">
    <children>
       <Button layoutX="-133.0" layoutY="135.0" mnemonicParsing="false" opacity="1.0" prefHeight="0.0" prefWidth="0.0" />
       <Label layoutX="207.0" layoutY="130.0" text="GRA">
