+++ Snake/src/sample/fxml/MenuController.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Font?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: BLACK;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.Controller.menuController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: BLACK;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.controller.MenuController">
    <children>
       <Label id="snakeTitle" fx:id="snakeLabel" layoutX="181.0" layoutY="14.0" prefHeight="125.0" prefWidth="239.0" style="-fx-font-family: &quot;Cyrulik Sharp&quot;;" text="SNAKE" textAlignment="CENTER" textFill="#8a2929">
          <font>
